New upstream version 5.30.0+dfsg

+++ src/calibre/gui2/preferences/coloring.py
@@ -13,17 +13,17 @@ from functools import partial
 from qt.core import (
     QAbstractItemView, QAbstractListModel, QApplication, QCheckBox, QComboBox,
     QDialog, QDialogButtonBox, QDoubleValidator, QFrame, QGridLayout, QIcon,
-    QIntValidator, QItemSelectionModel, QLabel, QLineEdit, QListView, QMenu,
+    QIntValidator, QItemSelectionModel, QLabel, QLineEdit, QListView,
     QPalette, QPushButton, QScrollArea, QSize, QSizePolicy, QSpacerItem,
     QStandardItem, QStandardItemModel, Qt, QToolButton, QVBoxLayout, QWidget,
-    QItemSelection, pyqtSignal
+    QItemSelection, QListWidget, QListWidgetItem, pyqtSignal
 )
 
 from calibre import as_unicode, prepare_string_for_xml, sanitize_file_name
 from calibre.constants import config_dir
 from calibre.gui2 import (
     choose_files, choose_save_file, error_dialog, gprefs, open_local_file,
-    pixmap_to_data
+    pixmap_to_data, question_dialog
 )
 from calibre.gui2.dialogs.template_dialog import TemplateDialog
 from calibre.gui2.metadata.single_download import RichTextDelegate
@@ -34,7 +34,7 @@ from calibre.library.coloring import (
 )
 from calibre.utils.icu import lower, sort_key
 from calibre.utils.localization import lang_map
-from polyglot.builtins import iteritems, unicode_type
+from polyglot.builtins import iteritems
 
 all_columns_string = _('All columns')
 
@@ -93,6 +93,12 @@ class ConditionEditor(QWidget):  # {{{
                 (_('is set'), 'is set'),
                 (_('is not set'), 'is not set'),
             ),
+            'multiple_no_isset' : (
+                (_('has'), 'has'),
+                (_('does not have'), 'does not have'),
+                (_('has pattern'), 'has pattern'),
+                (_('does not have pattern'), 'does not have pattern'),
+            ),
             'single'   : (
                 (_('is'), 'is'),
                 (_('is not'), 'is not'),
@@ -102,6 +108,14 @@ class ConditionEditor(QWidget):  # {{{
                 (_('does not match pattern'), 'does not match pattern'),
                 (_('is set'), 'is set'),
                 (_('is not set'), 'is not set'),
+            ),
+            'single_no_isset'   : (
+                (_('is'), 'is'),
+                (_('is not'), 'is not'),
+                (_('contains'), 'contains'),
+                (_('does not contain'), 'does not contain'),
+                (_('matches pattern'), 'matches pattern'),
+                (_('does not match pattern'), 'does not match pattern'),
             ),
     }
 
@@ -158,12 +172,12 @@ class ConditionEditor(QWidget):  # {{{
     @property
     def current_col(self):
         idx = self.column_box.currentIndex()
-        return unicode_type(self.column_box.itemData(idx) or '')
+        return str(self.column_box.itemData(idx) or '')
 
     @current_col.setter
     def current_col(self, val):
         for idx in range(self.column_box.count()):
-            c = unicode_type(self.column_box.itemData(idx) or '')
+            c = str(self.column_box.itemData(idx) or '')
             if c == val:
                 self.column_box.setCurrentIndex(idx)
                 return
@@ -172,12 +186,12 @@ class ConditionEditor(QWidget):  # {{{
     @property
     def current_action(self):
         idx = self.action_box.currentIndex()
-        return unicode_type(self.action_box.itemData(idx) or '')
+        return str(self.action_box.itemData(idx) or '')
 
     @current_action.setter
     def current_action(self, val):
         for idx in range(self.action_box.count()):
-            c = unicode_type(self.action_box.itemData(idx) or '')
+            c = str(self.action_box.itemData(idx) or '')
             if c == val:
                 self.action_box.setCurrentIndex(idx)
                 return
@@ -185,7 +199,7 @@ class ConditionEditor(QWidget):  # {{{
 
     @property
     def current_val(self):
-        ans = unicode_type(self.value_box.text()).strip()
+        ans = str(self.value_box.text()).strip()
         if self.current_col == 'languages':
             rmap = {lower(v):k for k, v in iteritems(lang_map())}
             ans = rmap.get(lower(ans), ans)
@@ -229,6 +243,10 @@ class ConditionEditor(QWidget):  # {{{
                     k = 'ondevice'
                 elif col == 'identifiers':
                     k = 'identifiers'
+                elif col == 'authors':
+                    k = 'multiple_no_isset'
+                elif col == 'title':
+                    k = 'single_no_isset'
                 else:
                     k = 'multiple' if m['is_multiple'] else 'single'
                 actions = self.action_map[k]
@@ -296,6 +314,52 @@ class ConditionEditor(QWidget):  # {{{
         if action in ('is set', 'is not set', 'is true', 'is false',
                 'is undefined'):
             self.value_box.setEnabled(False)
+# }}}
+
+
+class RemoveIconFileDialog(QDialog):  # {{{
+    def __init__(self, parent, icon_file_names, icon_folder):
+        self.files_to_remove = []
+        QDialog.__init__(self, parent)
+        self.setWindowTitle(_('Remove icons'))
+        self.setWindowFlags(self.windowFlags()&(~Qt.WindowType.WindowContextHelpButtonHint))
+        l = QVBoxLayout(self)
+        t = QLabel('<p>' + _('Select the icons you wish to remove. The icon files will be '
+                             'removed when you press OK. There is no undo.') + '</p>')
+        t.setWordWrap(True)
+        t.setTextFormat(Qt.TextFormat.RichText)
+        l.addWidget(t)
+        self.listbox = lw = QListWidget(parent)
+        lw.setSelectionMode(QAbstractItemView.SelectionMode.MultiSelection)
+        for fn in icon_file_names:
+            item = QListWidgetItem(fn)
+            item.setIcon(QIcon(os.path.join(icon_folder, fn)))
+            lw.addItem(item)
+        l.addWidget(lw)
+        self.bb = bb = QDialogButtonBox(
+            QDialogButtonBox.StandardButton.Ok|QDialogButtonBox.StandardButton.Cancel)
+        bb.accepted.connect(self.accept)
+        bb.rejected.connect(self.reject)
+        l.addWidget(bb)
+
+    def sizeHint(self):
+        return QSize(700, 600)
+
+    def accept(self):
+        self.files_to_remove = [item.text() for item in self.listbox.selectedItems()]
+        if not self.files_to_remove:
+            return error_dialog(self, _('No icons selected'), _(
+                'You must select at least one icon to remove'), show=True)
+        if question_dialog(self,
+            _('Remove icons'),
+            ngettext('One icon will be removed.', '{} icons will be removed.', len(self.files_to_remove)
+                        ).format(len(self.files_to_remove)) + ' ' + _('This will prevent any rules that use this icon from working. Are you sure?'),
+            yes_text=_('Yes'),
+            no_text=_('No'),
+            det_msg='\n'.join(self.files_to_remove),
+            skip_dialog_name='remove_icon_confirmation_dialog'
+        ):
+            QDialog.accept(self)
 # }}}
 
 
@@ -428,11 +492,11 @@ class RuleEditor(QDialog):  # {{{
         bb.rejected.connect(self.reject)
         l.addWidget(bb, 9, 0, 1, 8)
         if self.rule_kind != 'color':
-            self.remove_button = b = bb.addButton(_('&Remove icon'), QDialogButtonBox.ButtonRole.ActionRole)
+            self.remove_button = b = bb.addButton(_('&Remove icons'), QDialogButtonBox.ButtonRole.ActionRole)
             b.setIcon(QIcon(I('minus.png')))
-            b.setMenu(QMenu(b))
-            b.setToolTip('<p>' + _('Remove a previously added icon. Note that doing so will cause rules that use it to stop working.'))
-            self.update_remove_button()
+            b.clicked.connect(self.remove_icon_file_dialog)
+            b.setToolTip('<p>' + _('Remove previously added icons. Note that removing an '
+                                   'icon will cause rules that use it to stop working.') + '</p>')
 
         self.conditions_widget = QWidget(self)
         sa.setWidget(self.conditions_widget)
@@ -510,8 +574,8 @@ class RuleEditor(QDialog):  # {{{
 
     def update_color_label(self):
         pal = QApplication.palette()
-        bg1 = unicode_type(pal.color(QPalette.ColorRole.Base).name())
-        bg2 = unicode_type(pal.color(QPalette.ColorRole.AlternateBase).name())
+        bg1 = str(pal.color(QPalette.ColorRole.Base).name())
+        bg2 = str(pal.color(QPalette.ColorRole.AlternateBase).name())
         c = self.color_box.color
         self.color_label.setText('''
             <span style="color: {c}; background-color: {bg1}">&nbsp;{st}&nbsp;</span>
@@ -547,7 +611,6 @@ class RuleEditor(QDialog):  # {{{
                         import traceback
                         traceback.print_exc()
                     self.update_filename_box()
-                    self.update_remove_button()
                 if self.doing_multiple:
                     if icon_name not in self.rule_icon_files:
                         self.rule_icon_files.append(icon_name)
@@ -567,10 +630,10 @@ class RuleEditor(QDialog):  # {{{
             for i in range(1, model.rowCount()):
                 item = model.item(i, 0)
                 if item.checkState() == Qt.CheckState.Checked:
-                    fnames.append(lower(unicode_type(item.text())))
+                    fnames.append(lower(str(item.text())))
             fname = ' : '.join(fnames)
         else:
-            fname = lower(unicode_type(self.filename_box.currentText()))
+            fname = lower(str(self.filename_box.currentText()))
         return fname
 
     def update_icon_filenames_in_box(self):
@@ -589,23 +652,18 @@ class RuleEditor(QDialog):  # {{{
                         item = model.item(idx)
                         item.setCheckState(Qt.CheckState.Checked)
 
-    def update_remove_button(self):
-        m = self.remove_button.menu()
-        m.clear()
-        for name in self.icon_file_names:
-            ac = m.addAction(QIcon(os.path.join(self.icon_folder, name)), name)
-            connect_lambda(ac.triggered, self, lambda self: self.remove_image(self.sender().text()))
-
-    def remove_image(self, name):
-        try:
-            os.remove(os.path.join(self.icon_folder, name))
-        except EnvironmentError:
-            pass
-        else:
-            self.populate_icon_filenames()
-            self.update_remove_button()
-            self.update_filename_box()
-            self.update_icon_filenames_in_box()
+    def remove_icon_file_dialog(self):
+        d = RemoveIconFileDialog(self, self.icon_file_names, self.icon_folder)
+        if d.exec_() == QDialog.DialogCode.Accepted:
+            if len(d.files_to_remove) > 0:
+                for name in d.files_to_remove:
+                    try:
+                        os.remove(os.path.join(self.icon_folder, name))
+                    except OSError:
+                        pass
+                self.populate_icon_filenames()
+                self.update_filename_box()
+                self.update_icon_filenames_in_box()
 
     def add_blank_condition(self):
         c = ConditionEditor(self.fm, parent=self.conditions_widget)
@@ -629,7 +687,7 @@ class RuleEditor(QDialog):  # {{{
             self.update_icon_filenames_in_box()
 
         for i in range(self.column_box.count()):
-            c = unicode_type(self.column_box.itemData(i) or '')
+            c = str(self.column_box.itemData(i) or '')
             if col == c:
                 self.column_box.setCurrentIndex(i)
                 break
@@ -683,13 +741,13 @@ class RuleEditor(QDialog):  # {{{
         else:
             r.color = self.color_box.color
         idx = self.column_box.currentIndex()
-        col = unicode_type(self.column_box.itemData(idx) or '')
+        col = str(self.column_box.itemData(idx) or '')
         for c in self.conditions:
             condition = c.condition
             if condition is not None:
                 r.add_condition(*condition)
         if self.rule_kind == 'icon':
-            kind = unicode_type(self.kind_box.itemData(
+            kind = str(self.kind_box.itemData(
                                     self.kind_box.currentIndex()) or '')
         else:
             kind = self.rule_kind
